% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_areas.R
\docType{data}
\name{DE_areas}
\alias{DE_areas}
\title{Shapefile
Shapefile from Germany Study Areas BfN}
\format{
Shapefile
}
\usage{
data(DE_areas)
}
\description{
Shapefile
Shapefile from Germany Study Areas BfN
}
\examples{
data(DE_areas)
}
\references{
Bundesamt für Naturschutz (BfN)
}
\keyword{datasets}
