% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bath.R
\docType{data}
\name{DE_bath}
\alias{DE_bath}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(DE_bath)
}
\description{
Shapefile from Germany Bathymetry
}
\examples{
data(DE_bath)
}
\references{
To-do
}
\keyword{datasets}
