% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_coast.R
\docType{data}
\name{DE_coast}
\alias{DE_coast}
\title{Shapefile}
\format{
Shapefile
}
\usage{
data(DE_coast)
}
\description{
Shapefile from Germany Coast
}
\examples{
data(DE_coast)
}
\references{
To-do
}
\keyword{datasets}
