% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_distancecoast.R
\docType{data}
\name{DE_distancecoast}
\alias{DE_distancecoast}
\title{Shapefile
Shapefile from German Distance to Coast}
\format{
Shapefile
}
\usage{
data(DE_distancecoast)
}
\description{
Shapefile
Shapefile from German Distance to Coast
}
\examples{
data(DE_distancecoast)
}
\references{
To-do
}
\keyword{datasets}
