% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_EEZ.R
\docType{data}
\name{DE_EEZ}
\alias{DE_EEZ}
\title{Shapefile
Shapefile from Germany EEZ}
\format{
Shapefile
}
\usage{
data(DE_EEZ)
}
\description{
Shapefile
Shapefile from Germany EEZ
}
\examples{
data(DE_EEZ)
}
\references{
To-do
}
\keyword{datasets}
