% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_grid10x10_EEZ.R
\docType{data}
\name{DE_grid10x10_EEZ}
\alias{DE_grid10x10_EEZ}
\title{Contains a sf object with a grid from the North Sea of 10 x 10
A total of 583 observations of 5 variables}
\format{
A grid 10x10 of the north sea
}
\usage{
data(DE_grid10x10_EEZ)
}
\description{
Contains a sf object with a grid from the North Sea of 10 x 10
A total of 583 observations of 5 variables
}
\examples{
data(DE_grid10x10_EEZ)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
