% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_ICES.R
\docType{data}
\name{DE_ICES}
\alias{DE_ICES}
\title{Shapefile
Shapefile from Germany ICES rectangles}
\format{
Shapefile
}
\usage{
data(DE_ICES)
}
\description{
Shapefile
Shapefile from Germany ICES rectangles
}
\examples{
data(DE_ICES)
}
\references{
To-do
}
\keyword{datasets}
