% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_land.R, R/Europe_land.R
\docType{data}
\name{DE_land}
\alias{DE_land}
\title{Shapefile
Shapefile from Germany land contour}
\format{
Shapefile

Shapefile
}
\usage{
data(DE_land)

data(DE_land)
}
\description{
Shapefile
Shapefile from Germany land contour

Shapefile
Shapefile from Europe land
}
\examples{
data(DE_land)
data(DE_land)
}
\references{
DIVA-GIS

To-do
}
\keyword{datasets}
