% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_natura.R
\docType{data}
\name{DE_natura}
\alias{DE_natura}
\title{Shapefile
Shapefile from Germany natura}
\format{
Shapefile
}
\usage{
data(DE_natura)
}
\description{
Shapefile
Shapefile from Germany natura
}
\examples{
data(DE_natura)
}
\references{
To-do
}
\keyword{datasets}
