% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_OWF.R
\docType{data}
\name{DE_OWF}
\alias{DE_OWF}
\title{Shapefile
Shapefile from Germany's Offshore Wind Farms}
\format{
Shapefile
}
\usage{
data(DE_OWF)
}
\description{
Shapefile
Shapefile from Germany's Offshore Wind Farms
}
\examples{
data(DE_OWF)
}
\references{
To-do
}
\keyword{datasets}
