% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_SCA.R
\docType{data}
\name{DE_SCA}
\alias{DE_SCA}
\title{Shapefile
Shapefile from German Special Conservation Areas}
\format{
Shapefile
}
\usage{
data(DE_SCA)
}
\description{
Shapefile
Shapefile from German Special Conservation Areas
}
\examples{
data(DE_SCA)
}
\references{
To-do
}
\keyword{datasets}
