% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_shipping.R
\docType{data}
\name{DE_shipping}
\alias{DE_shipping}
\title{Shapefile
Shapefile from DE_shipping}
\format{
Shapefile
}
\usage{
data(DE_shipping)
}
\description{
Shapefile
Shapefile from DE_shipping
}
\examples{
data(DE_shipping)
}
\references{
To-do
}
\keyword{datasets}
